refactor: move modal components into Screen folder
The separation of modals and screens into separate folders didn't make
much sense and made it harder for newcomers to know how to organise
things. From now on modal components should have a suffix in the filename e.g.
`ShortcutsModal` to make them easier to find.
