+++ ui/Screen/Wallet/Pool/CollectModal.svelte
