+++ ui/Screen/Wallet/Pool/PoolOnboardingModal/Erc20Allowance.svelte
