+++ ui/Screen/Wallet/Pool/PoolOnboardingModal/Intro.svelte
