+++ ui/Screen/Wallet/Pool/PoolOnboardingModal/Review.svelte
