+++ ui/Screen/Wallet/QrCodeModal.svelte
