+++ ui/src/customProtocolHandler.ts
@@ -10,7 +10,7 @@ import * as modal from "./modal";
 import * as error from "./error";
 import { inputStore } from "./search";
 
-import ModalSearch from "../Modal/Search.svelte";
+import SearchModal from "ui/Screen/SearchModal.svelte";
 
 export const register = (): void => {
   ipc.listenCustomProtocolInvocation(async message => {
@@ -75,5 +75,5 @@ const handleMessage = (message: ipc.CustomProtocolInvocation): void => {
   }
 
   inputStore.set(urn);
-  modal.show(ModalSearch);
+  modal.show(SearchModal);
 };
